New option for input_uvc.so: -e/--every_frame
When applied input_uvc will drop all frames but the one stated, eg, when set to 10, frames 1-9 will be read and ignored and frame 10 will be processed as usual.
This option is iseful with crappy devices that produce noise when signal is not read fast enough (which may happen if you use slow processors that fail to compress images at a decent speed).
I use this option with a noname EasyCap clone made in China and old Raspberry Pi.